Various changes to address bugs in hierarchies
The targeted bugs were introduced when we switched to the possibility of having multiple instances of ClassHierarchyProvider (and for other entity types) per project to allow different rooted hierarchies.  The ChangeManager assumed a project scoped singleton ClassHierarchyProvider (as did other various actions).  This meant that changes to the hierarchy would sometimes be propagated and at other times would not be propagated, depending on how a hierarchy provider was obtained.  To fix this the ChangeManager now uses the HierarchyProviderManager to propagate changes to all managed hierarchies rather than the project singleton hierarchies for each type of entities.

We (currently) will still allow injecting of ClassHierarchyProvider etc. for project scoped beans as this is a big change.  To accommodate this the various HierarchyProviderFactory implementations e.g. ClassHierarchyProviderFactory will check the roots to see if the project singleton can be returned (for roots that are the standard roots in an ontology e.g. owl:Thing for classes, owl:topObjectProperty for object properties etc.) and if so the project singleton hierarchy provider will be returned.  This ensures there is an effective project singleton hierarchy provider for each entity type.  At some point we should no longer allow the injection of ClassHierarchyProvider etc. instead the HierarchyProviderManager should be injected and the hierarchy provider should be retrieved from this.